attepmt to fit data with b-splines
setup loads the voltage clamp and simulates the states and current output. Cubic B-splines used for fitting data in two ways to ensure the good fit around the voltage changes: non-uniform grid that has 6 knots around each switching point, and a set of B-splines with piecewise support that is collected into a single collocation matrix.