Siimple model estimation working
try_ask_tell_simple_model.py - a working version of the estimator that tracks the dynamics of a single gating variable.

try_segments.py - attempt at segmenting B-spline space to fit various intervals separately. Currently the segments share a point at the end/start but the width of splines is different on either side. Need to stitch things properly by either using an overlapping surface or defining a spline of the same width at the start and end of segments.